worked on listview and listview builder
